+++ changes.md
@@ -3,4 +3,4 @@
 
 ### 9 August 2025
 * `getTokenRate` now returns one token unit rate, without required `weiAmount`
-* `tokenRateUSD` added to `loadPayments` returned `paymentsData` array
+* `tokenRateUSD`, `paidPriceUSD`, `paidTotalUSD`, `paidFeeUSD` added to `loadPayments` returned `paymentsData` array